Partitioned `storage` module
