+++ src/lib.rs
@@ -1,6 +1,5 @@
 extern crate chrono;
 
-pub mod container;
 pub mod errors;
 pub mod device;
 pub mod io;
